requirements.txt
lxml
beautifulsoup4
requests
python-telegram-bot